refactor(frontend/src/shared/ui/Button/Button.jsx):
move Button.jsx to shared/ui/Button for

reusable UI components
